Refactor marshallers
